Fix service wait queue
