+++ src/examples/analysis/GridZonal.cpp
@@ -40,7 +40,7 @@ int main(int argc, char* argv[])
   terrama2::core::registerFactories();
 
   auto& serviceManager = terrama2::core::ServiceManager::getInstance();
-te::core::URI uri("pgsql://"+TERRAMA2_DATABASE_USERNAME+":"+TERRAMA2_DATABASE_PASSWORD+"@"+TERRAMA2_DATABASE_HOST+":"+TERRAMA2_DATABASE_PORT+"/"+TERRAMA2_DATABASE_DBNAME);
+  te::core::URI uri("pgsql://"+TERRAMA2_DATABASE_USERNAME+":"+TERRAMA2_DATABASE_PASSWORD+"@"+TERRAMA2_DATABASE_HOST+":"+TERRAMA2_DATABASE_PORT+"/"+TERRAMA2_DATABASE_DBNAME);
   serviceManager.setLogConnectionInfo(uri);
 
   terrama2::services::analysis::core::PythonInterpreterInit pythonInterpreterInit;
