+++ src/terrama2/services/appserver/main.cpp
@@ -215,7 +215,7 @@ int main(int argc, char* argv[])
       // Must initialize the python interpreter before creating any thread.
       terrama2::services::analysis::core::PythonInterpreterInit pythonInterpreterInit;
 
-      QCoreApplication app(argc, argv);
+      QGuiApplication app(argc, argv);
 
       // Changes in the initialization order may cause locale problems
       std::locale::global(std::locale::classic());
